#3 minor refactoring
