+++ src/main/kotlin/com/github/christophpickl/payconiq/service/exceptions.kt
@@ -9,20 +9,3 @@ import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExcep
 import mu.KotlinLogging.logger as klogger
 
 class NotFoundException(message: String) : RuntimeException(message)
-
-@ControllerAdvice
-@RestController
-class CustomExceptionHandler : ResponseEntityExceptionHandler() {
-
-    private val log = klogger {}
-
-    @ExceptionHandler(NotFoundException::class)
-    fun handleNotFoundException(exception: NotFoundException): ResponseEntity<ApiError> {
-        log.warn(exception) { "Not found handled." }
-        return ResponseEntity(ApiError("Not found"), HttpStatus.NOT_FOUND)
-    }
-}
-
-data class ApiError(
-    val message: String
-)
